@@ -14,8 +14,7 @@
     "gulp": "^3.9.1",
     "jest-expo": "^22.0.0",
     "json-loader": "^0.5.7",
-    "less": "^2.7.3",
-    "less-loader": "^4.0.5",
+    "node-sass": "^4.7.2",
     "react-native-scripts": "1.7.0",
     "react-test-renderer": "16.0.0-beta.5",
     "react-transform-catch-errors": "^1.0.2",
@@ -26,9 +25,9 @@
     "redux-devtools-dock-monitor": "^1.1.2",
     "redux-devtools-log-monitor": "^1.4.0",
     "redux-logger": "^3.0.6",
+    "sass-loader": "^6.0.6",
     "style-loader": "^0.19.0",
-    "webpack": "^3.8.1",
-    "webpack-dev-middleware": "^1.12.0",
+    "webpack": "^3.10.0",
     "webpack-dev-server": "^2.9.4",
     "webpack-hot-middleware": "^2.20.0"
   },
@@ -50,7 +49,6 @@
     "babel": "^6.23.0",
     "babel-eslint": "^8.0.2",
     "babel-polyfill": "^6.26.0",
-    "clone": "^2.1.1",
     "eslint": "^4.11.0",
     "eslint-config-airbnb": "^16.1.0",
     "eslint-plugin-import": "^2.8.0",
