@@ -1,11 +1,12 @@
-/* global __DEV__ */
+const devMode = (process.env.NODE_ENV !== 'development');
+
 export default {
   // App Details
   appName: 'Starter Kit',
 
   // Build Configuration - eg. Debug or Release?
-  DEV: __DEV__,
+  DEV: devMode,
 
   // Google Analytics - uses a 'dev' account while we're testing
-  gaTrackingId: (__DEV__) ? 'UA-84284256-2' : 'UA-84284256-1',
+  gaTrackingId: (devMode) ? 'UA-84284256-2' : 'UA-84284256-1',
 };
