@@ -1,5 +1,4 @@
-import React, { Component } from 'react';
-import PropTypes from 'prop-types';
+import React from 'react';
 import { View } from 'react-native';
 import { List, ListItem } from 'react-native-elements';
 import { Actions } from 'react-native-router-flux';
@@ -10,20 +9,21 @@ const listItemProps = {
   },
 };
 
-export default class Menu extends Component {
-  render = () => (
-    <View
-      style={{
-        flex: 1,
-        backgroundColor: '#F3F5F5',
-        flexDirection: 'column',
-      }}
-    >
-      <List containerStyle={{ marginTop: 44 }}>
-        <ListItem {...listItemProps} key={'0'} title={'Home'} />
-        <ListItem {...listItemProps} key={'1'} title={'About Us'} onPress={() => Actions.home()} />
-        <ListItem {...listItemProps} key={'2'} title={'Settings'} />
-      </List>
-    </View>
-  )
-}
+const Menu = () => (
+  <View
+    style={{
+      flex: 1,
+      paddingTop: 44,
+      backgroundColor: '#F3F5F5',
+      flexDirection: 'column',
+    }}
+  >
+    <List>
+      <ListItem {...listItemProps} key="0" title="Home" />
+      <ListItem {...listItemProps} key="1" title="About Us" onPress={() => Actions.home()} />
+      <ListItem {...listItemProps} key="2" title="Settings" />
+    </List>
+  </View>
+);
+
+export default Menu;
