@@ -1,37 +1,22 @@
-import clone from 'clone';
-import assign from 'object-assign';
+import { combineReducers } from 'redux';
 
-const initialState = {
-  color: 'red',
-  data: {
-    loading: false,
-    objects: [],
-  },
+// Our custom reducers
+// We need to import each one here and add them to the combiner at the bottom
+import colors from './colors';
+import member from './member';
+import recipes from './recipes';
+
+// Combine all
+const appReducer = combineReducers({
+  colors,
+  member,
+  recipes,
+});
+
+// Setup root reducer
+const rootReducer = (state, action) => {
+  const newState = (action.type === 'RESET') ? undefined : state;
+  return appReducer(newState, action);
 };
 
-export default function reduce(state = initialState, action) {
-  switch (action.type) {
-  case 'TOGGLE_COLOR':
-    return assign({}, state, {
-      color: state.color === 'red' ? 'blue' : 'red'
-    });
-
-  case 'EXAMPLE_REQUEST_START':
-    return assign({}, state, {
-      data: assign({}, state.data, {
-        loading: true,
-      }),
-    });
-
-  case 'EXAMPLE_REQUEST_DATA':
-    return assign({}, state, {
-      data: assign({}, state.data, {
-        loading: false,
-        objects: action.data,
-      }),
-    });
-
-  default:
-    return state;
-  }
-}
+export default rootReducer;
